+++ src/custom-error.ts
@@ -1,3 +1,5 @@
+import { GraphQLError } from 'graphql';
+
 export class CustomError extends Error {
   code: number;
   additionalInfo?: string;
@@ -8,3 +10,12 @@ export class CustomError extends Error {
     this.additionalInfo = additionalInfo;
   }
 }
+
+export const formatError = (error: GraphQLError) => {
+  const formattedError = error.originalError as CustomError;
+  return {
+    code: formattedError.code,
+    message: formattedError.message,
+    additionalInfo: formattedError?.additionalInfo,
+  };
+};
