+++ src/resolvers.ts
@@ -1,6 +1,5 @@
 import { User } from './user';
 import { dataSource } from './data-source';
-import { GraphQLError } from 'graphql';
 import { CustomError } from './custom-error';
 import { validateStrongPassword } from './input-validation';
 import bcrypt from 'bcrypt';
@@ -21,7 +20,7 @@ export const resolvers = {
   Mutation: {
     createUser: async (_parent: never, args: { data: UserInput }) => {
       if (!validateStrongPassword(args.data.password)) {
-        throw new CustomError(404, 'erro: senha fraca');
+        throw new CustomError(400, 'Senha fraca', 'Deve ter 6 caracteres, com no mínimo 1 letra e 1 dígito');
       }
 
       const existingUser = await dataSource.getRepository(User).findOne({
@@ -31,7 +30,7 @@ export const resolvers = {
       });
 
       if (existingUser) {
-        throw new GraphQLError('erro: email utilizado');
+        throw new CustomError(409, 'Email já registrado');
       }
 
       const user = new User();
