+++ src/setup.ts
@@ -2,6 +2,7 @@ import { DataSource } from 'typeorm';
 import { ApolloServer } from 'apollo-server';
 import { typeDefs } from './schema';
 import { resolvers } from './resolvers';
+import { formatError } from './custom-error';
 
 export const initializeDbConnection = async (dataSource: DataSource) => {
   await dataSource
@@ -16,6 +17,7 @@ export const launchServer = async () => {
   const server = new ApolloServer({
     typeDefs,
     resolvers,
+    formatError,
   });
 
   const port = 3000;
