+++ src/test/index.ts
@@ -12,3 +12,6 @@ before(async () => {
 
 require('./hello')
 require('./create-user');
+require('./duplicate-email');
+require('./weak-password');
+
